[Move `@kbn/config-schema` to server] `data`
