+++ src/plugins/data/public/search/session/sessions_mgmt/components/main.test.tsx
@@ -20,7 +20,7 @@ import { LocaleWrapper } from '../__mocks__';
 import { SearchSessionsMgmtMain } from './main';
 import { SharePluginStart } from '@kbn/share-plugin/public';
 import { sharePluginMock } from '@kbn/share-plugin/public/mocks';
-import { SearchSessionsConfigSchema } from '../../../../../config';
+import type { SearchSessionsConfigSchema } from '../../../../../server/config';
 import { createSearchUsageCollectorMock } from '../../../collectors/mocks';
 
 let mockCoreSetup: MockedKeys<CoreSetup>;
