+++ src/plugins/data/public/search/session/sessions_mgmt/lib/get_columns.tsx
@@ -30,7 +30,7 @@ import { SearchSessionsMgmtAPI } from './api';
 import { getExpirationStatus } from './get_expiration_status';
 import { UISession } from '../types';
 import { SearchUsageCollector } from '../../../collectors';
-import { SearchSessionsConfigSchema } from '../../../../../config';
+import type { SearchSessionsConfigSchema } from '../../../../../server/config';
 
 // Helper function: translate an app string to EuiIcon-friendly string
 const appToIcon = (app: string) => {
