+++ src/plugins/data/public/search/session/sessions_mgmt/lib/get_expiration_status.ts
@@ -8,7 +8,7 @@
 
 import { i18n } from '@kbn/i18n';
 import moment from 'moment';
-import { SearchSessionsConfigSchema } from '../../../../../config';
+import type { SearchSessionsConfigSchema } from '../../../../../server/config';
 
 export const getExpirationStatus = (config: SearchSessionsConfigSchema, expires: string | null) => {
   const tNow = moment.utc().valueOf();
