+++ src/plugins/data/server/search/session/get_session_status.test.ts
@@ -10,7 +10,7 @@ import { elasticsearchServiceMock } from '@kbn/core/server/mocks';
 import { getSessionStatus } from './get_session_status';
 import { SearchSessionSavedObjectAttributes, SearchSessionStatus } from '../../../common';
 import moment from 'moment';
-import { SearchSessionsConfigSchema } from '../../../config';
+import { SearchSessionsConfigSchema } from '../../config';
 
 const mockInProgressSearchResponse = {
   body: {
