+++ src/plugins/data/server/search/session/get_session_status.ts
@@ -10,7 +10,7 @@ import moment from 'moment';
 import { ElasticsearchClient } from '@kbn/core/server';
 import { SearchSessionSavedObjectAttributes, SearchSessionStatus } from '../../../common';
 import { SearchStatus } from './types';
-import { SearchSessionsConfigSchema } from '../../../config';
+import { SearchSessionsConfigSchema } from '../../config';
 import { getSearchStatus } from './get_search_status';
 
 export async function getSessionStatus(
