+++ src/plugins/data/server/search/strategies/common/async_utils.ts
@@ -11,7 +11,7 @@ import {
   AsyncSearchGetRequest,
 } from '@elastic/elasticsearch/lib/api/types';
 import { ISearchOptions } from '@kbn/search-types';
-import { SearchConfigSchema } from '../../../../config';
+import { SearchConfigSchema } from '../../../config';
 
 /**
  @internal
