+++ src/plugins/data/server/search/strategies/esql_async_search/esql_async_search_strategy.ts
@@ -22,7 +22,7 @@ import { getKbnSearchError } from '../../report_search_error';
 import type { ISearchStrategy, SearchStrategyDependencies } from '../../types';
 import type { IAsyncSearchOptions } from '../../../../common';
 import { toAsyncKibanaSearchResponse } from './response_utils';
-import { SearchConfigSchema } from '../../../../config';
+import { SearchConfigSchema } from '../../../config';
 
 // `drop_null_columns` is going to change the response
 // now we get `all_columns` and `columns`
